feat: create separate Role model file
- Create dedicated backend/app/database/models/role.py
- Implement Role model with hierarchical support and permissions
- Add fields: id, name, code, description, permissions (JSON), parent_id
- Include audit timestamps (created_at, updated_at)
- Support role hierarchy through self-referencing parent_id
- Remove Role model from user.py to separate concerns
- Update imports in repositories, services, and __init__.py
- Maintain backward compatibility with existing schemas

Task #73: Implement Role Model, Schemas, and Repository
