+++ backend/app/database/models/__init__.py
@@ -1,4 +1,5 @@
 from .base import Base
-from .user import User, Department, Stage, Role, UserSupervisor
+from .user import User, Department, Stage, UserSupervisor
+from .role import Role
 
 __all__ = ["Base", "User", "Department", "Stage", "Role", "UserSupervisor"]
