+++ backend/app/services/role_service.py
@@ -2,7 +2,7 @@ from typing import List
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy import select
 
-from ..database.models.user import Role
+from ..database.models.role import Role
 from ..schemas.user import Role as RoleSchema
